Consolidate cc shortcuts
